Start new python SNMP library, called snapy
This is based on pynetsnmp but without the pieces I don't want and won't
try in any way to be compatible with twistedsnmp. The approach to
handling requests will be different, with the intent to be reliable.

Also a major goal is to develop a comprehensive unit test suite along
side this new library so I can be confident that it won't suck.