Rewrite SNMP walk, use GETBULK if available.
This new implementation of walk should behave similar to the snmpwalk
and snmpbulkwalk commands, the former for v1, the later for v2. The main
differences between this code and those commands are that multiple OIDs
are accepted and a simple GET on the given OIDs before the walk starts
rather than at the end.

Note that this includes an API change: the netsnmp callbacks now report
the results as a list of tuples rather than dicts. This is important
because when EndOfMibView is reported it uses the same OID as the last
value, thus a dict could only report one or the other but not both.