Don't combine SNMP v1 queries.
Without GETBULK we don't get any advantage to doing walks on multiple
OIDs. We could still combine GET queries but that probably would
complicate the code without that much gain. This of course depends on
what turns out to be the most reliable way of talking to our more
fragile v1 servers such as Oracle.