Various fixes for the snmp walk implementation.
Net-SNMP doesn't appear to handle non_repeaters=0 as I expected, I'm not
sure if this is a bug or I just don't understand non_repeaters. In any
case we need to make smaller requests so the response packets are
smaller. I'm seeing lots of dropped packets somewhere when snmpd's
default max size of 1500 is reached.