Rework NagiosWriter to properly unregister itself.
Previously the NagiosWriter simply added itself as a writer during start
up and then forever stayed in the event loop. This was problematic
because the reactor will then constantly call doWrite() as long as the
pipe is writable which is most of the time. The result pegs the cpu.

Also in the process tweak its logging to be less noisy. Since generally
the reason for failed writes is simply someone restarting Nagios it
isn't helpful to fill the logs with errors and tracebacks. It will now
only re-try the open every 10 seconds, logging a warning if it fails.