Do not impose a limit on nagnet's command queue.
The new NagiosWriter class imposed a limit on the number of commands it
would queue to ensure that nagcat wouldn't eat memory if nagios was down
for an extended period of time. This caused problems for nagnet which
may submit thousands of downtime commands all at once so this limit
should only be enforced in nagcat, not nagnet.