Add support for using snmp display hints.
This adds some infrastructure for getting at data provided by MIBs.
Although it would have been simpler to just use the more generic
snprint_value so we don't have to bother with knowing anything about
MIBs but that would force all strings to be ascii. When there is no hint
provided we still want to pass along the raw string which may be
anything. The OID class as been extended to provide this info.

The result parsing functions have been updated to use the hint.

TODO: it would be also good to start using enum values as well.