Rework the notify module to break it into plugins.
Although this doesn't change functionality much it does aid the
development of notification methods that cannot go into the public
nagcat repository (yet). It would be cute to also rework query.py into
plugins as well but that can wait till there is a need.

Note that the notify script's arguments have changed a bit.