Use an alternate threadpool for oracle queries.
When a connection fails during initialization in just the right way it
will hang forever. The current cx_Oracle provides no way to time it out.
The current work around to keep nagcat running is to do this in a thread
so at least only that test will hang and stop running, not everything.
However during shutdown twisted tries to gracefully stop its threads
which has the end result of making nagcat hang forever if a thread is
hung. To work around this issue I'm adding a new threadpool that marks
all of its threads as daemon threads and never bothers to try to stop
them gracefully. It ain't pretty, but this is oracle so that is a given.