Add a generic check_freshness Nagios check plugin.
This test is intended for running as a Nagios check plugin and will
report any checks that have gone stale in the status file. This is
useful for reporting some quick detail on any tests that may be stuck in
nagcat. This is preferable to the freshness support in nagios itself
because if nagcat itself is down only this test would alert, not every
test which is a bit friendlier.