Run oracle queries in subprocesses rather than threads.
When a call to cx_Oracle hangs during connect() there is no (safe) way
to kill it if it is in a thread. There was a suggested a trick with
signals but I couldn't get that to work. Killing with pthread_cancel()
the thread does successfully kill the call but it leaves OCI's locking
in an inconsistent state so future calls may deadlock. The only safe way
I've found so far to use Oracle is to push the code into a subprocess
which can be SIGTERMed to kill it.

This implementation hooks into Twisted's Process class in weird ways to
run a piece of python code rather than execing an entirely new program.
The result is pickled and sent back to nagcat over a pipe.