Add new warning_time_limit attribute for tests.
This provides a simple way to cause a test to jump from warning to
critical if it has been in a warning state for too long. I previously
tried to implement this in a generic way with the nagios status query
support but nagcat's logic limitations prevent that from being used for
this purpose and I haven't found a solution. So here is a non-generic
way to do it with the extra bonus that it is way simpler than any
generic method would have been.