Add preliminary support for parsing performance data.
An obvious extension to running Nagios plugins is directly parsing
performance data to record it into RRD and generally simplify the
parsing done within tests themselves.

This is imperfect as it does not yet combine queries properly but I
wanted to checkpoint progress as this at least seems to work.