Account for nagcat test repeat intervals in freshness check.
If a test's repeat interval is longer than the thresholds of this check
fails alerts will come up. This test does its best to grab that repeat
interval without fully expanding the nagcat config which would be slow.
Chances are the repeat value isn't a Link or include a variable
reference so this works for me for now. Something more sophisticated
would be required if that is ever not true.