Real fix this time: Only group tasks together with the same repeat interval.
If you load the following tests, all sharing a common query:
    1. Register Test1 with repeat 3600s, creates new group.
    2. Register Test2 with repeat 60s, creates new group (60 < 3600)
    3. Register Test3 with repeat 3600s...

When that third test loads the fun happens:
    4. The two groups for Test1 and Test2 are found.
    5. One is popped out of the set, could be either.
       To trigger the bug we need to pop the 3600s group.
    6. The 60s group is now merged into the 3600s group.

And poof, now Test2 will only get scheduled once an hour because it
moved over to the wrong group at a later step. The fix is to simply only
consider groups with the exact same repeat interval so mismatched groups
are not eligible for merging.

Note: This change bumps the minimum Python version, chances are < 2.6
probably doesn't work anymore anyway and I know this change needs 2.6.